requirements.txt
PyYAML==6.0.2
asyncpg==0.30.0
httpx==0.28.1
numpy==1.26.4
openvino-telemetry==2025.2.0
openvino==2025.2.0
orjson==3.10.15
pandas==2.2.3
prefect==3.4.10
psutil==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
starlette==0.41.3
typing-inspection==0.4.1
opencv-python==4.9.0.80
tqdm
psycopg[binary]==3.2.2